Add safety check for the cfme version in the repo file.
The repo file (for example
http://file.cloudforms.lab.eng.rdu2.redhat.com/builds/cfme/5.10/5.10.3.1/packages.repo
) contains link to baseurl with dir `latest`. But not every time we have
picked the latest package.repo. In such case the package refered in
`latest` dir will not be the intended one and it is better to detect
that. This patch adds a method for checking the version of cfme the
packages.repo in the end ponts at.